Test/unit tests backend (#26)
* feat: Se agregaron las pruebas unitarias del router del modulo de jugadores

Pruebas unitarias de los modulos del repository
Se realizo un cambio minimo en models para que permita paremotros de los IDs
Cambio en repository para normalizar el item que se busca por ser "case-sensitive matching"

* feat: Pruebas unitarias del router de players

Se agregaron las pruebas unitarias de cada uno de los modulos del router de players.

* fix: Se modularizaron las funciones para los test

Se pusieron las funciones en un archivo llamado factories.py donde solo se tiene que llamar para facilitar la reutilzacion de codigo

* feat: Se agregaron las pruebas unitarias de archivo de validators del modulo de los jugadores

Se agregaron las funciones de test para los modulos del archivo validators
Se arreglaron los comentarios
Se quitaron todas las advertencias que habian cuando se corrian las pruebas unitarias

+++ backend/src/modules/teams/models.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, Integer, BigInteger, String, Boolean, DateTime, func, ForeignKey, Index, text
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, func, ForeignKey, Index, text
 from config.database import Base
 
 class Team(Base):
     __tablename__ = "teams"
 
-    id = Column(BigInteger, primary_key=True, index=True)
+    id = Column(Integer, primary_key=True, index=True)
     name = Column(String(128), nullable=False)         # DB has CITEXT UNIQUE(name)
     city = Column(String(128), nullable=False)
     image_url = Column(String(512), nullable=True)
